Add files via upload
The agricultural sector can be considered the backbone of any developing economy. In order to achieve maximum crop yield, it is necessary to provide farmers with the best technologies and methodologies. Artificial intelligence can be of great help in the control of crop diseases. In recent years, thanks to the advantages of machine learning and feature extraction, it has been widely concerned by academia and industry. At the same time, it has also become a research hot spot in the field of agricultural plant protection. Various diseases have affected the natural growth of plants, and infected plants are the main factors of plant production loss. Manual detection and identification of plant diseases require careful and observant examination by expertise. To overcome the manual examination procedures, automated identification and detection can be involved which provide faster, scalable and accurate solutions. In this research, our convolutional neural network based system and the pre-trained ResNet50 model was tested on the Plant Pathology 2021 - FGVC8 database organized at kaggle. The experiments showed that the model based on ResNet50 is more efficient than the first two models based on convolutional neural networks in terms of accuracy rate (91.83%). The comparison of the results found showed that the number of epochs, the depth of networks, and the size of the image base are important factors for obtaining better results.